+++ getiryemek-clone/src/main/java/com/example/getiryemek_clone/repository/AddressRepository.java
@@ -3,5 +3,6 @@ package com.example.getiryemek_clone.repository;
 import com.example.getiryemek_clone.entity.Address;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AddressRepository extends JpaRepository<Address,Integer> {
+public interface AddressRepository extends JpaRepository<Address,Long> {
+
 }
