+++ getiryemek-clone/src/main/java/com/example/getiryemek_clone/repository/RestaurantAdminRepository.java
@@ -3,5 +3,5 @@ package com.example.getiryemek_clone.repository;
 import com.example.getiryemek_clone.entity.RestaurantAdmin;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RestaurantAdminRepository extends JpaRepository<RestaurantAdmin,Integer> {
+public interface RestaurantAdminRepository extends JpaRepository<RestaurantAdmin,Long> {
 }
